requirements.txt
fastapi
uvicorn
httpx
python-dotenv
numpy
scipy
geopy
pydantic
haversine
discord.py
